Consolidate all OSM classes in one file
